Add a middleware to allow whitelisted routes
Rails has several middleware that allocate resources to handle requests.
Since the routing middleware is loaded last, that means resources are
allocated even when the routes are invalid, which leaves applications
vulnerable to DOS attacks.

This middleware allows specified routes to be served, but rejects any
that are not. If inserted before middleware that allocate resources, it
will respond to invalid requests before they resources are allocated.