Lazily merge config options in the middleware
Since the routes are loaded in an after initialize hook and the
middlewares aren't in the railtie, routes have yet to be configured when
the options has is merged with the config. This causes the route
matchers to be empty and all requests to be blocked (since no routes
have been configured according to the options hash).

Instead, lazily merge the config options with the config so that we can
allow the railtie to configure the routes in Rails apps.