Make allow middleware is thread-safe
Rails' middleware stack is not reinitialized with each request by Puma
(per http://eftimov.net/writing-rails-middleware). We were storing the
Rack env in an instance variable and referencing it later, which is
perfectly safe if you are guaranteed to always have a fresh instance of
the middleware. It turns out that the same instance is used for all
requests, so it was not at all safe.

Temporarily dup the middleware stack until we can rework this with a
more permanent fix.