Return 404 HEAD response by default
Update the Allow middleware to return a 404 HEAD response (with empty
headers and body), and add config options for :content, :content_type,
and :status. If :content is nil (the default), no headers will be
returned. Otherwise, a content-type header will be added. If
:content_type is nil (the default), "text/plain" will be used.

With this change, :message is no longer needed and has been removed (use
:content instead).